:adhedive_band: feat : git 충돌 해결
