+++ src/components/Profile/ProfileSection/ProfileFollow/styledFollow.jsx
@@ -1,0 +1,14 @@
+import { Link } from "react-router-dom";
+import styled from "styled-components";
+
+const Anchor = styled(Link)`
+  text-decoration: none;
+  text-align: center;
+  margin: auto;
+
+  strong {
+    font-size: 18px;
+  }
+`;
+
+export { Anchor };
