+++ src/components/common/SetProfile/SetProfile.jsx
@@ -1,28 +1,14 @@
-import React from "react";
-import styled from "styled-components";
-import { Button } from "../../../GlobalStyle";
+import { Form, StartButton } from "./styledSetProfileInput";
 import SetprofileHeader from "./SetProfileHeader/SetprofileHeader";
 import SetProfileImage from "./SetProfileImage/SetProfileImage";
 import SetProfileInput from "./SetProfileInput/SetProfileInput";
 
-const Wrapper = styled.div`
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-`;
-
-const StartButton = styled(Button)`
-  width: 80%;
-  font-size: 16px;
-  color: white;
-  padding: 13px 0;
-  margin: 20px 0;
-`;
-
 export default function SetProfile() {
+  const submitHandler = (event) => {
+    event.preventDefault();
+  };
   return (
-    <Wrapper>
+    <Form>
       <SetprofileHeader />
       <SetProfileImage />
       <SetProfileInput
@@ -40,7 +26,7 @@ export default function SetProfile() {
         label="소개"
         placeholder="자신과 판매할 상품에 대해 소개해 주세요!"
       />
-      <StartButton>감귤마켓 시작하기</StartButton>
-    </Wrapper>
+      <StartButton onClick={submitHandler}>감귤마켓 시작하기</StartButton>
+    </Form>
   );
 }
