+++ src/pages/Profile/ProfileEdit/ProfileEdit.jsx
@@ -1,0 +1,9 @@
+import SetProfile from "../../../components/common/SetProfile/SetProfile";
+
+export default function ProfileEdit() {
+  return (
+    <>
+      <SetProfile />
+    </>
+  );
+}
